% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GACppEigen}
\alias{GACppEigen}
\title{GA (Eigen)}
\usage{
GACppEigen(population_size, bitstring_size, lower, upper,
  maximum_number_of_iterations, tolerance, maximum_number_of_iterations_equal)
}
\arguments{
\item{population_size}{The size of the population.}

\item{bitstring_size}{The size of the individual.}

\item{lower}{The lower bound of the interval.}

\item{upper}{The upper bound of the interval.}

\item{maximum_number_of_iterations}{The maximum allowed number of interations.}

\item{tolerance}{The convergence tolerance.}

\item{maximum_number_of_iterations_equal}{The number of iterations without changing the fittest individual before forced convergence.}
}
\value{
A list with two elements: the fittest decoded individual found in the entire run and its fitness.
}
\description{
Simple genetic algorithm for maximising function. Implemented using the boost and Eigen libraries.
}
