% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGA.R
\docType{package}
\name{simpleGA}
\alias{simpleGA}
\alias{simpleGA-package}
\title{simpleGA}
\description{
A short and simple implementation of a genetic algorithm for maximising univariate functions.
}
\author{
S. B. Vilsen
}
